Documentation of internals

Documentation of internals

Modules

Constants

Types

Functions

Macros

Index

Modules

PredictMD.PredictMD — Module

PredictMD.Cleaning — Module

PredictMD.Compilation — Module

PredictMD.GPU — Module

PredictMD.Server — Module

Constants

Types

PredictMD.CrossValidation — Type

PredictMD.DataFrameFeatureContrasts — Type

PredictMD.DataFrameFeatureContrasts — Method

PredictMD.DecisionTreeModel — Type

PredictMD.GLMModel — Type

PredictMD.ImmutableDataFrame2GLMSingleLabelBinaryClassTransformer — Type

PredictMD.ImmutableFeatureArrayTransposerTransformer — Type

PredictMD.ImmutablePackageMultiLabelPredictionTransformer — Type

PredictMD.ImmutablePackageSingleLabelPredictProbaTransformer — Type

PredictMD.ImmutablePackageSingleLabelPredictionTransformer — Type

PredictMD.ImmutablePredictProbaSingleLabelInt2StringTransformer — Type

PredictMD.ImmutablePredictionsSingleLabelInt2StringTransformer — Type

PredictMD.KnetModel — Type

PredictMD.LIBSVMModel — Type

PredictMD.MutableDataFrame2ClassificationKnetTransformer — Type

PredictMD.MutableDataFrame2DecisionTreeTransformer — Type

PredictMD.MutableDataFrame2RegressionKnetTransformer — Type

PredictMD.SimplePipeline — Type

PredictMD.SimplePipeline — Method

Functions

PredictMD.DataFrame2LIBSVMTransformer — Method

PredictMD.accuracy — Method

PredictMD.auprc — Method

PredictMD.aurocc — Method

PredictMD.averageprecisionscore — Method

PredictMD.avg_precision — Method

PredictMD.binary_brier_score — Method

binary_brier_score(ytrue, yscore)

Computes the binary formulation of the Brier score, defined as:

\[\frac{1}{N}\sum\limits _{t=1}^{N}(f_t-o_t)^2 \,\!\]

Lower values are better. Best value is 0.

PredictMD.calculate_smote_pct_under — Method

PredictMD.check_column_types — Method

PredictMD.cohen_kappa — Method

PredictMD.cohen_kappa — Method

PredictMD.cohen_kappa — Method

PredictMD.compute_contingency_table — Method

PredictMD.compute_contingency_table — Method

PredictMD.convert_value_to_missing! — Function

PredictMD.delete_nothings! — Method

PredictMD.f1score — Method

PredictMD.false_negative_rate — Method

PredictMD.false_positive_rate — Method

PredictMD.fbetascore — Method

PredictMD.filename_extension — Method

PredictMD.fit! — Function

PredictMD.fit! — Function

PredictMD.fit! — Function

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fit! — Method

PredictMD.fix_column_types! — Method

PredictMD.fix_type — Function

PredictMD.flatten — Method

PredictMD.generate_feature_contrasts — Method

PredictMD.generate_formula — Method

PredictMD.generate_formula — Method

PredictMD.generate_formula — Method

PredictMD.get_binary_thresholds — Method

PredictMD.get_history — Method

PredictMD.get_history — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.get_underlying — Method

PredictMD.getallrocnums — Method

PredictMD.getlabelint2stringmap — Method

PredictMD.getlabelstring2intmap — Method

PredictMD.icd9_code_to_single_level_dx_ccs — Method

PredictMD.inverse — Method

PredictMD.is_ci — Function

PredictMD.is_ci_or_runtests — Function

PredictMD.is_ci_or_runtests_or_docs_or_examples — Function

PredictMD.is_deploy_docs — Function

PredictMD.is_docs_or_examples — Function

PredictMD.is_make_docs — Function

PredictMD.is_make_examples — Function

PredictMD.is_nothing — Function

PredictMD.is_one_to_one — Method

PredictMD.is_runtests — Function

PredictMD.is_square — Method

PredictMD.is_travis_ci — Function

PredictMD.is_travis_ci_on_linux — Function

PredictMD.isflat — Function

PredictMD.ispipeline — Function

PredictMD.load_model — Method

PredictMD.mean_square_error — Method

mean_square_error(ytrue, ypred)

PredictMD.multilabelprobabilitiestopredictions — Method

PredictMD.multilabelprobabilitiestopredictions — Method

PredictMD.negative_predictive_value — Method

PredictMD.open_plots_during_tests — Function

PredictMD.package_directory — Method

package_directory(parts...)::String

Equivalent to abspath(joinpath(abspath(package_directory()), parts...)).

PredictMD.package_directory — Method

package_directory()::String

Return the PredictMD package directory.

PredictMD.parse_functions! — Method

PredictMD.plot_probability_calibration_curve — Method

PredictMD.plot_probability_calibration_curve — Method

PredictMD.plotlearningcurve — Function

PredictMD.plotlearningcurve — Function

PredictMD.plotlearningcurve — Method

PredictMD.plotprcurve — Method

PredictMD.plotprcurve — Method

PredictMD.plotroccurve — Method

PredictMD.plotroccurve — Method

PredictMD.plotsinglelabelbinaryclassifierhistogram — Method

PredictMD.plotsinglelabelregressiontrueversuspredicted — Method

PredictMD.positive_predictive_value — Method

PredictMD.prcurve — Method

PredictMD.prcurve — Method

PredictMD.precision — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predict_proba — Method

PredictMD.predictionsassoctodataframe — Function

PredictMD.probability_calibration_metrics — Function

PredictMD.probability_calibration_metrics — Method

PredictMD.probability_calibration_scores_and_fractions — Method

PredictMD.probability_calibration_scores_and_fractions — Method

PredictMD.r2_score — Method

r2_score(ytrue, ypred)

Computes coefficient of determination. Higher values are better. Best value is 1.

PredictMD.recall — Method

PredictMD.risk_score_cutoff_values — Method

PredictMD.risk_score_cutoff_values — Method

PredictMD.roccurve — Method

PredictMD.roccurve — Method

PredictMD.root_mean_square_error — Method

root_mean_square_error(ytrue, ypred)

PredictMD.save_model — Method

PredictMD.sensitivity — Method

PredictMD.set_feature_contrasts! — Method

PredictMD.set_feature_contrasts! — Method

PredictMD.set_feature_contrasts! — Method

PredictMD.set_feature_contrasts! — Method

PredictMD.set_max_epochs! — Method

PredictMD.set_max_epochs! — Method

PredictMD.shuffle_rows! — Method

PredictMD.shuffle_rows! — Method

PredictMD.simple_linear_regression — Method

simple_linear_regression(x::AbstractVector, y::AbstractVector)

Simple linear regression - given a set of two-dimensional points (x, y), use the ordinary least squares method to find the best fit line of the form y = a + b*x (where a and b are real numbers) and return the tuple (a, b).

PredictMD.simple_moving_average — Method

PredictMD.single_labeldataframeknetregression — Method

PredictMD.single_labeldataframeknetregression_Knet — Method

PredictMD.single_labeldataframelinearregression — Method

PredictMD.single_labeldataframelinearregression_GLM — Method

PredictMD.single_labeldataframerandomforestregression — Method

PredictMD.single_labeldataframerandomforestregression_DecisionTree — Method

PredictMD.single_labeldataframesvmregression — Method

PredictMD.single_labeldataframesvmregression_LIBSVM — Method

PredictMD.single_labelmulticlassdataframeknetclassifier — Method

PredictMD.single_labelmulticlassdataframeknetclassifier_Knet — Method

PredictMD.single_labelmulticlassdataframerandomforestclassifier — Method

PredictMD.single_labelmulticlassdataframesvmclassifier — Method

PredictMD.single_labelmulticlassdataframesvmclassifier_LIBSVM — Method

PredictMD.single_labelmulticlassdfrandomforestclassifier_DecisionTree — Method

PredictMD.single_labelprobabilitiestopredictions — Method

PredictMD.singlelabelbinaryclassdataframelogisticclassifier — Method

PredictMD.singlelabelbinaryclassdataframelogisticclassifier_GLM — Method

PredictMD.singlelabelbinaryclassificationmetrics — Method

PredictMD.singlelabelbinaryclassificationmetrics — Method

PredictMD.singlelabelbinaryclassificationmetrics_resultdict — Method

PredictMD.singlelabelbinaryclassificationmetrics_tunableparam — Method

PredictMD.singlelabelbinaryyscore — Method

PredictMD.singlelabelbinaryytrue — Method

PredictMD.singlelabelregressionmetrics — Method

PredictMD.singlelabelregressionmetrics — Method

PredictMD.singlelabelregressionmetrics_resultdict — Method

PredictMD.singlelabelregressionypred — Method

PredictMD.singlelabelregressionytrue — Method

PredictMD.smote — Method

PredictMD.smote — Method

PredictMD.specificity — Method

PredictMD.split_data — Method

PredictMD.split_data — Method

PredictMD.transform — Function

PredictMD.transform — Function

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform — Method

PredictMD.transform_columns! — Function

PredictMD.trapz — Function

trapz(x, y)

Compute the area under the curve of 2-dimensional points (x, y) using the trapezoidal method.

PredictMD.true_negative_rate — Method

PredictMD.true_positive_rate — Method

PredictMD.tuplify — Function

PredictMD.version — Method

version()::VersionNumber

Return the version number of PredictMD.

PredictMD.Cleaning.ccs_onehot_names — Function

Given a dataframe, return the column names corresponding to CCS "one-hot" columns.

Examples

import CSVFiles
import FileIO
import PredictMD

df = DataFrames.DataFrame(
    FileIO.load(
        MY_CSV_FILE_NAME;
        type_detect_rows = 30_000,
        )
    )

@info(PredictMD.Cleaning.ccs_onehot_names(df))
@info(PredictMD.Cleaning.ccs_onehot_names(df, "ccs_onehot_"))

PredictMD.Cleaning.clean_hcup_nis_csv_icd9 — Method

Given a single ICD 9 code, import the relevant patients from the Health Care Utilization Project (HCUP) National Inpatient Sample (NIS) database.

Examples:

import CSVFiles
import FileIO
import PredictMD

icd_code_list = ["8841"]
icd_code_type=:procedure
input_file_name_list = [
    "./data/nis_2012_core.csv",
    "./data/nis_2013_core.csv",
    "./data/nis_2014_core.csv",
    ]
output_file_name = "./output/hcup_nis_pr_8841.csv"

PredictMD.Cleaning.clean_hcup_nis_csv_icd9(
    icd_code_list,
    input_file_name_list,
    output_file_name;
    icd_code_type=icd_code_type,
    rows_for_type_detect = 30_000,
    )

df = DataFrames.DataFrame(
    FileIO.load(
        output_file_name;
        type_detect_rows = 30_000,
        )
    )

@info(PredictMD.Cleaning.ccs_onehot_names(df))

PredictMD.Cleaning.column_names_with_prefix — Method

PredictMD.Cleaning.symbol_begins_with — Method

PredictMD.Cleaning.x_contains_y — Method

Macros

Index

PredictMD.Cleaning

PredictMD.Compilation

PredictMD.GPU

PredictMD.PredictMD

PredictMD.Server

PredictMD.CrossValidation

PredictMD.DataFrameFeatureContrasts

PredictMD.DataFrameFeatureContrasts

PredictMD.DecisionTreeModel

PredictMD.GLMModel

PredictMD.ImmutableDataFrame2GLMSingleLabelBinaryClassTransformer

PredictMD.ImmutableFeatureArrayTransposerTransformer

PredictMD.ImmutablePackageMultiLabelPredictionTransformer

PredictMD.ImmutablePackageSingleLabelPredictProbaTransformer

PredictMD.ImmutablePackageSingleLabelPredictionTransformer

PredictMD.ImmutablePredictProbaSingleLabelInt2StringTransformer

PredictMD.ImmutablePredictionsSingleLabelInt2StringTransformer

PredictMD.KnetModel

PredictMD.LIBSVMModel

PredictMD.MutableDataFrame2ClassificationKnetTransformer

PredictMD.MutableDataFrame2DecisionTreeTransformer

PredictMD.MutableDataFrame2RegressionKnetTransformer

PredictMD.SimplePipeline

PredictMD.SimplePipeline

PredictMD.Cleaning.ccs_onehot_names

PredictMD.Cleaning.clean_hcup_nis_csv_icd9

PredictMD.Cleaning.column_names_with_prefix

PredictMD.Cleaning.symbol_begins_with

PredictMD.Cleaning.x_contains_y

PredictMD.DataFrame2LIBSVMTransformer

PredictMD.accuracy

PredictMD.auprc

PredictMD.aurocc

PredictMD.averageprecisionscore

PredictMD.avg_precision

PredictMD.binary_brier_score

PredictMD.calculate_smote_pct_under

PredictMD.check_column_types

PredictMD.cohen_kappa

PredictMD.cohen_kappa

PredictMD.cohen_kappa

PredictMD.compute_contingency_table

PredictMD.compute_contingency_table

PredictMD.convert_value_to_missing!

PredictMD.delete_nothings!

PredictMD.f1score

PredictMD.false_negative_rate

PredictMD.false_positive_rate

PredictMD.fbetascore

PredictMD.filename_extension

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fit!

PredictMD.fix_column_types!

PredictMD.fix_type

PredictMD.flatten

PredictMD.generate_feature_contrasts

PredictMD.generate_formula

PredictMD.generate_formula

PredictMD.generate_formula

PredictMD.get_binary_thresholds

PredictMD.get_history

PredictMD.get_history

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.get_underlying

PredictMD.getallrocnums

PredictMD.getlabelint2stringmap

PredictMD.getlabelstring2intmap

PredictMD.icd9_code_to_single_level_dx_ccs

PredictMD.inverse

PredictMD.is_ci

PredictMD.is_ci_or_runtests

PredictMD.is_ci_or_runtests_or_docs_or_examples

PredictMD.is_deploy_docs

PredictMD.is_docs_or_examples

PredictMD.is_make_docs

PredictMD.is_make_examples

PredictMD.is_nothing

PredictMD.is_one_to_one

PredictMD.is_runtests

PredictMD.is_square

PredictMD.is_travis_ci

PredictMD.is_travis_ci_on_linux

PredictMD.isflat

PredictMD.ispipeline

PredictMD.load_model

PredictMD.mean_square_error

PredictMD.multilabelprobabilitiestopredictions

PredictMD.multilabelprobabilitiestopredictions

PredictMD.negative_predictive_value

PredictMD.open_plots_during_tests

PredictMD.package_directory

PredictMD.package_directory

PredictMD.parse_functions!

PredictMD.plot_probability_calibration_curve

PredictMD.plot_probability_calibration_curve

PredictMD.plotlearningcurve

PredictMD.plotlearningcurve

PredictMD.plotlearningcurve

PredictMD.plotprcurve

PredictMD.plotprcurve

PredictMD.plotroccurve

PredictMD.plotroccurve

PredictMD.plotsinglelabelbinaryclassifierhistogram

PredictMD.plotsinglelabelregressiontrueversuspredicted

PredictMD.positive_predictive_value

PredictMD.prcurve

PredictMD.prcurve

PredictMD.precision

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predict_proba

PredictMD.predictionsassoctodataframe

PredictMD.probability_calibration_metrics

PredictMD.probability_calibration_metrics

PredictMD.probability_calibration_scores_and_fractions

PredictMD.probability_calibration_scores_and_fractions

PredictMD.r2_score

PredictMD.recall

PredictMD.risk_score_cutoff_values

PredictMD.risk_score_cutoff_values

PredictMD.roccurve

PredictMD.roccurve

PredictMD.root_mean_square_error

PredictMD.save_model

PredictMD.sensitivity

PredictMD.set_feature_contrasts!

PredictMD.set_feature_contrasts!

PredictMD.set_feature_contrasts!

PredictMD.set_feature_contrasts!

PredictMD.set_max_epochs!

PredictMD.set_max_epochs!

PredictMD.shuffle_rows!

PredictMD.shuffle_rows!

PredictMD.simple_linear_regression

PredictMD.simple_moving_average

PredictMD.single_labeldataframeknetregression

PredictMD.single_labeldataframeknetregression_Knet

PredictMD.single_labeldataframelinearregression

PredictMD.single_labeldataframelinearregression_GLM

PredictMD.single_labeldataframerandomforestregression

PredictMD.single_labeldataframerandomforestregression_DecisionTree

PredictMD.single_labeldataframesvmregression

PredictMD.single_labeldataframesvmregression_LIBSVM

PredictMD.single_labelmulticlassdataframeknetclassifier

PredictMD.single_labelmulticlassdataframeknetclassifier_Knet

PredictMD.single_labelmulticlassdataframerandomforestclassifier

PredictMD.single_labelmulticlassdataframesvmclassifier

PredictMD.single_labelmulticlassdataframesvmclassifier_LIBSVM

PredictMD.single_labelmulticlassdfrandomforestclassifier_DecisionTree

PredictMD.single_labelprobabilitiestopredictions

PredictMD.singlelabelbinaryclassdataframelogisticclassifier

PredictMD.singlelabelbinaryclassdataframelogisticclassifier_GLM

PredictMD.singlelabelbinaryclassificationmetrics

PredictMD.singlelabelbinaryclassificationmetrics

PredictMD.singlelabelbinaryclassificationmetrics_resultdict

PredictMD.singlelabelbinaryclassificationmetrics_tunableparam

PredictMD.singlelabelbinaryyscore

PredictMD.singlelabelbinaryytrue

PredictMD.singlelabelregressionmetrics

PredictMD.singlelabelregressionmetrics

PredictMD.singlelabelregressionmetrics_resultdict

PredictMD.singlelabelregressionypred

PredictMD.singlelabelregressionytrue

PredictMD.smote

PredictMD.smote

PredictMD.specificity

PredictMD.split_data

PredictMD.split_data

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform

PredictMD.transform_columns!

PredictMD.trapz

PredictMD.true_negative_rate

PredictMD.true_positive_rate

PredictMD.tuplify

PredictMD.version